# spanlogger: fix double logging of caller information