--- a/spanlogger/spanlogger_test.go
+++ b/spanlogger/spanlogger_test.go
@@ -7,6 +7,7 @@ import (
 	"io"
 	"path/filepath"
 	"runtime"
+	"slices"
 	"strings"
 	"testing"
 
@@ -206,7 +207,7 @@ func BenchmarkSpanLoggerWithRealLogger(b *testing.B) {
 		b.Run(name, func(b *testing.B) {
 			buf := bytes.NewBuffer(nil)
 			logger := dskit_log.NewGoKitWithWriter("logfmt", buf)
-			logger = log.With(logger, "ts", log.DefaultTimestampUTC, "caller", log.Caller(5))
+			logger = log.With(logger, "ts", log.DefaultTimestampUTC, "caller", SpanLoggerAwareCaller(5))
 
 			if debugEnabled {
 				logger = level.NewFilter(logger, level.AllowAll())
@@ -267,12 +268,12 @@ type loggerWithDebugEnabled struct {
 
 func (l loggerWithDebugEnabled) DebugEnabled() bool { return l.debugEnabled }
 
-func TestSpanLogger_CallerInfo(t *testing.T) {
+func TestSpanLoggerAwareCaller(t *testing.T) {
 	testCases := map[string]func(w io.Writer) log.Logger{
 		// This is based on Mimir's default logging configuration: https://github.com/grafana/mimir/blob/50d1c27b4ad82b265ff5a865345bec2d726f64ef/pkg/util/log/log.go#L45-L46
 		"default logger": func(w io.Writer) log.Logger {
 			logger := dskit_log.NewGoKitWithWriter("logfmt", w)
-			logger = log.With(logger, "ts", log.DefaultTimestampUTC, "caller", log.Caller(5))
+			logger = log.With(logger, "ts", log.DefaultTimestampUTC, "caller", SpanLoggerAwareCaller(5))
 			logger = level.NewFilter(logger, level.AllowAll())
 			return logger
 		},
@@ -280,7 +281,7 @@ func TestSpanLogger_CallerInfo(t *testing.T) {
 		// This is based on Mimir's logging configuration with rate-limiting enabled: https://github.com/grafana/mimir/blob/50d1c27b4ad82b265ff5a865345bec2d726f64ef/pkg/util/log/log.go#L42-L43
 		"rate-limited logger": func(w io.Writer) log.Logger {
 			logger := dskit_log.NewGoKitWithWriter("logfmt", w)
-			logger = log.With(logger, "ts", log.DefaultTimestampUTC, "caller", log.Caller(6))
+			logger = log.With(logger, "ts", log.DefaultTimestampUTC, "caller", SpanLoggerAwareCaller(6))
 			logger = dskit_log.NewRateLimitedLogger(logger, 1000, 1000, nil)
 			logger = level.NewFilter(logger, level.AllowAll())
 			return logger
@@ -288,7 +289,7 @@ func TestSpanLogger_CallerInfo(t *testing.T) {
 
 		"default logger that has been wrapped with further information": func(w io.Writer) log.Logger {
 			logger := dskit_log.NewGoKitWithWriter("logfmt", w)
-			logger = log.With(logger, "ts", log.DefaultTimestampUTC, "caller", log.Caller(5))
+			logger = log.With(logger, "ts", log.DefaultTimestampUTC, "caller", SpanLoggerAwareCaller(5))
 			logger = level.NewFilter(logger, level.AllowAll())
 			logger = log.With(logger, "user", "user-1")
 			return logger