@charleskorn 2024-10-09
I'd like to investigate and understand the regression in the `SpanLoggerWithRealLogger/all_levels_allowed/level.Debug` benchmark, but I'm going to share this for feedback on the approach first.