@charleskorn 2024-10-10
> I'd like to investigate and understand the regression in the `SpanLoggerWithRealLogger/all_levels_allowed/level.Debug` benchmark, but I'm going to share this for feedback on the approach first.

I've investigated this and addressed it in the main PR description. This PR is good to go now.