@aknuds1 2024-10-10
> This is unavoidable, applications that are sensitive to this regression could use SpanLogger.DebugLog() rather than wrapping SpanLogger with level.Debug(...).

@charleskorn does this regression only apply to debug level?