@charleskorn 2024-10-10
> > This is unavoidable, applications that are sensitive to this regression could use SpanLogger.DebugLog() rather than wrapping SpanLogger with level.Debug(...).
> 
> @charleskorn does this regression only apply to debug level?

It applies to other levels too, if they are used when wrapping a `SpanLogger`. 

However, at least in Mimir, it seems that the vast majority of logging to a `SpanLogger` is at the debug level, so that the logs appear only as events on trace spans and not in logs running at the default info level, so I don't think this regression is a big issue.