@aknuds1 2024-10-11
> However, at least in Mimir, it seems that the vast majority of logging to a SpanLogger is at the debug level, so that the logs appear only as events on trace spans and not in logs running at the default info level, so I don't think this regression is a big issue.

Ah good point.